Downloaded library to toggle/show password

+++ twodeesapp/app/Http/Controllers/ProfileController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use App\Models\Location;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 
 class ProfileController extends Controller {
     public function index($id) {
@@ -40,9 +41,8 @@ class ProfileController extends Controller {
                 'profile_picture' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048',
             ]);
 
-            $profile_picture = $request->file('profile_picture');
             $profile_picture_name = time() . '.' . $request->profile_picture->extension();
-            $profile_picture->move(public_path('profile_pictures'), $profile_picture_name);
+            $request->profile_picture->move(public_path('profile_pictures'), $profile_picture_name);
             
             $user->profile_picture = 'profile_pictures/' . $profile_picture_name;
             $user->save();
@@ -57,7 +57,7 @@ class ProfileController extends Controller {
         $user = User::find($id);
 
         if($user != null) {
-            if(file_exists(public_path($user->profile_picture))) {
+            if($user->profile_picture && file_exists(public_path($user->profile_picture))) {
                 unlink(public_path($user->profile_picture));
             }
 
@@ -89,13 +89,12 @@ class ProfileController extends Controller {
                 'profile_picture' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048',
             ]);
 
-            if(file_exists(public_path($user->profile_picture))) {
+            if($user->profile_picture && file_exists(public_path($user->profile_picture))) {
                 unlink(public_path($user->profile_picture));
             }
 
-            $profile_picture = $request->file('profile_picture');
             $profile_picture_name = time() . '.' . $request->profile_picture->extension();
-            $profile_picture->move(public_path('profile_pictures'), $profile_picture_name);
+            $request->profile_picture->move(public_path('profile_pictures'), $profile_picture_name);
             
             $user->profile_picture = 'profile_pictures/' . $profile_picture_name;
             $user->save();
@@ -112,6 +111,49 @@ class ProfileController extends Controller {
 
         if($user != null) {
             return view('profile.edit', compact('user', 'locations'));
+        } else {
+            return redirect()->route('/')->with('error', 'The profile you are searching for does not exist!');
+        }
+    }
+
+    public function update($id, Request $request) {
+        $user = User::find($id);
+
+        if($user != null) {
+            $request->validate([
+                'name' => 'required',
+                'email' => 'required|email|unique:users,email,' . $id,
+                'username' => 'required|max:255|unique:users,username,' . $id,
+                'birth_date' => 'required|date|before:' .now()->subYears(18)->addDays(1)->toDateString(),
+                'location_id' => 'required|exists:locations,id',
+                'profile_picture' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048'
+            ]);
+
+            if ($request->filled('password')) {
+                $user->password = Hash::make($request->password);
+            }
+
+            if($request->hasFile('profile_picture')) 
+            {
+                if($user->profile_picture && file_exists(public_path($user->profile_picture))) {
+                    unlink(public_path($user->profile_picture));
+                }
+
+                $profile_picture_name = time() . '.' . $request->profile_picture->extension();
+                $request->profile_picture->move(public_path('profile_pictures'), $profile_picture_name);
+                $user->profile_picture = 'profile_pictures/' . $profile_picture_name;
+            }
+
+            $user->name = $request->name;
+            $user->email = $request->email;
+            $user->username = $request->username;
+            $user->birth_date = $request->birth_date;
+            $user->password = bcrypt(value: $user->password);
+            $user->location_id = $request->location_id;
+            $user->bio = trim($request->bio);
+            $user->save();
+
+            return redirect()->route('profile.index', $user->id)->with('success', 'Your profile has been updated!');
         } else {
             return redirect()->route('/')->with('error', 'The profile you are searching for does not exist!');
         }
